Assignments 4.13-4.23

+++ osa4/controllers/blogs.js
@@ -1,35 +1,43 @@
 const blogRouter = require("express").Router()
 const Blog = require("../models/blog")
+const userExtractor = require("../utils/middleware").userExtractor
+
 
 blogRouter.get("/", async (request, response) => {
   const blogs = await Blog.find({})
+    .populate("user", {username: 1, name: 1, id: 1})
 
   response.json(blogs)
-
-  /*Blog
-    .find({})
-    .then(blogs => {
-      response.json(blogs)
-    })*/
 })
 
-blogRouter.post("/", async (request, response) => {
-  const blog = new Blog(request.body)
+blogRouter.post("/", userExtractor, async (request, response) => {
+  const body = request.body
+  const user = request.user
+
+  const blog = new Blog ({
+    title: body.title,
+    url: body.url,
+    likes: body.likes,
+    user: user._id
+  })
 
   const result = await blog.save()
+  user.blogs = user.blogs.concat(result._id)
+  await user.save()
+
   response.status(201).json(result)
-  
-  /*blog
-    .save()
-    .then(result => {
-      response.status(201).json(result)
-    })*/
 })
 
-blogRouter.delete("/:id", async (request, response) => {
-  console.log(request.params.id)
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+blogRouter.delete("/:id", userExtractor, async (request, response) => {
+  const user = request.user
+  const blog = await Blog.findById(request.params.id)
+
+  if (user._id.toString() === blog.user.toString()){
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  } else {
+    response.status(401).json({ error: "invalid token" })
+  }
 })
 
 blogRouter.put("/:id", async(request, response) => {
